requirements.txt
Flask
requests
azure-identity
azure-mgmt-resource